refactor server init code
